requirements.txt
torch
torchvision
gradio
scikit-learn
pillow
